Improve and centralize page override functionality
